virtiofsd: Open O_WRONLY files as O_RDWR
Guest can open a file O_WRONLY and later do some writes. In that case
guest will ask daemon to mmap(PROT_WRITE) in dax memory window that that
fails. Looks like mmap(PROT_WRITE) can succeed only if file has been
opended both with read and write permissions (O_RDWR). So to take care
of this case, upgrade to O_RDWR.

I see that there is code to do that if user has mounted with option
-o writeback. I don't understand that option fully yet, so I am not
relying on that yet. May be later, we need to merge this two or
create another option for virtio-fs only.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>